+++ js/tamarin/core/E4XNode.h
@@ -305,7 +305,7 @@ namespace avmplus
 		// If the low bit of this integer is set, this value points directly
 		// to a single child (one E4XNode *).  If there are multiple children,
 		// this points to an AtomArray containing E4XNode pointers (NOT Atoms)
-		DWB(int) m_children;
+		DWB(uintptr) m_children;
 		#define SINGLECHILDBIT 0x1
 
 		friend class E4XNode;
